+++ EduCDM/MIRT/MIRT.py
@@ -167,8 +167,8 @@ class MIRT(CDM):
             users.extend(user_id.tolist())
 
         metric.update(torch.tensor(y_pred), torch.tensor(y_true))
-        y_pred_rounded = torch.round(torch.tensor(y_pred))
-        y_true_rounded = torch.round(torch.tensor(y_true))
+        y_pred_rounded = torch.round(torch.tensor(y_pred)).int()
+        y_true_rounded = torch.round(torch.tensor(y_true)).int()
         precision.update(y_pred_rounded, y_true_rounded)
         recall.update(y_pred_rounded, y_true_rounded)
         f1.update(y_pred_rounded,y_true_rounded)
